requirements.txt
openai>=1.50.0
gradio>=5.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
